# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "(Quiz) Server API - Swagger",
    "description": "Account Admin: admin - 12345 <br /> <br /> Account User: user - 12345",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "definitions": {
...
    "AnswerDTO": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "NodeJS is a JavaScript runtime built on Chrome's V8 JavaScript engine?"
        },
        "isCorrect": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "AnswerCreateDTO": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "NodeJS is a JavaScript runtime built on Chrome's V8 JavaScript engine?"
        },
        "isCorrect": {
          "type": "boolean",
          "example": true
        },
        "question": {
          "type": "object",
          "example": {
            "id": 1
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/answer/get-all": {
      "get": {
        "tags": ["Answer Controller"],
        "summary": "Get the list of answer",
        "operationId": "getAllAnswer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully - List of answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/answer/get-one/{id}": {
      "get": {
        "tags": ["Answer Controller"],
        "summary": "Get answer by id",
        "produces": ["application/xml", "application/json"],
        "operationId": "getAnswerById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of answer to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully - Answer Info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID! ID Must be a number"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Answer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/answer/create": {
      "post": {
        "tags": ["Answer Controller"],
        "summary": "Create answer",
        "operationId": "createAnswer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "AnswerDTO",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/AnswerCreateDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/definitions/AnswerCreateDTO"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/definitions/AnswerCreateDTO"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/definitions/AnswerCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/AnswerCreateDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/AnswerCreateDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/AnswerCreateDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/AnswerCreateDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/answer/update": {
      "put": {
        "tags": ["Answer Controller"],
        "summary": "Update answer",
        "operationId": "updateAnswer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "AnswerDTO",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/AnswerDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/definitions/AnswerDTO"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/definitions/AnswerDTO"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/definitions/AnswerDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Answer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/answer/delete/{id}": {
      "delete": {
        "tags": ["Answer Controller"],
        "summary": "Delete answer",
        "produces": ["application/xml", "application/json"],
        "operationId": "deleteAnswerById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of answer to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/definitions/AnswerDTO"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Answer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}